+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2024-04-23
+### Added
+- verbose logging will print processing time of a file
+### Fixed
+- compressing files over 2GB
+
 ## [2.0.0] - 2024-01-25
 ### Breaking
 - compression is now done using Node's native zlib module, this requires Node v12 at least